+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Deprecate Aqua (QA) as a target IntelliJ Platform
 - Print requested IntelliJ Platform when throwing exception about unresolved dependency.
 - Make `IntelliJPlatformDependenciesHelper` aware of custom IntelliJ Platform when used within custom tasks.
+- Move the Coroutines JavaAgent lock file into module's build directory
 
 ### Fixed